Added comments
Error is fixed! Code should run smoothly now (It did not like inserting into a set from a vector). I suggest you copy and paste into a new project then edit. Code is computing all of additions that add up to said values in inputVector and all sub additions. This algorithm gets rid of duplicates hence the checkDuplicates( ) use. (exception of pair (1,1) for some reason this one duplicates in certain scenarios. If this could be fixed it would be greatly appreciated XD.

All that is left TODO is compute the most efficient path while keeping track of it and outputting it.

After all of the above is done file input should be implemented. (Hence why I commented it out, it is better to just use inputVector while implementing functionality).